pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "symmetric"
version = "1.0"
authors = [
    { name = "Eugen Hoppe", email = "rispe_keller0x@icloud.com" }
]
description = "Symmetric is a streamlined Python library for symmetric encryption"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
urls = { "Homepage" = "https://github.com/eugen-hoppe/symmetric" }

dependencies = [
    "cryptography==42.0.5",
    "python-dotenv==1.0.1"
]

[project.scripts]
stringkeys = "src.symmetric:main"

[project.optional-dependencies]

[tool.setuptools]
package-dir = { "" = "src" }
packages = {"find" = { "where" = ["src"], "include" = ["*"] }}
include-package-data = true
